cql3 unit test: add assertion for the number of serialized columns
The result sets that the assertions are performed against
are result sets before serialization to the user and therefore
contain also columns that will not be serialized and sent as
the query's final result. The patch adds an assertion on the
number of columns that will be present in the final serialized
result.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>