auth: Add resiliancy against potentialy corrupted table
There has been a case where reading from the internal
authentication roles table resulted in a memorry access
violation. The problem was that some columns that was assumed
to have value for each row of the table didn't. According to
the code this situation shouldn't have happened, but it did.
This change adds a check for existance of those columns (can_login
and superuser columns). In the case they don't exist an error
is issued but the defult value for those columns is used.
The reason is that the error will not self correct by crashing and
in a lot of cases those fields are not relevant. One case that
can still happen is that if the corrupte row is indeed the superuser
that can loggin, the administrator will be left out of the system.
A manual correction of the table will be needed which is the case
even before our patch (the system will crash instead of just no logging
the user in), however, after this patch the log will indicate the
problem so we will know it needs fixing by changing the table manually
and from outside scylla.

Fixes #4226