auth: Prevent race between role_manager and pasword_authenticator
When scylla is started for the first time with PasswordAuthenticator
enabled, there it can be that a record of the default superuser
will be created in the table with the can_login and is_superuser
set to null. This happens because the module in charge of creating
the row is the role manger and the module in charge of setting the
default password is the password authenticator.
Those two modules are started together, it the case when the
password authenticator finish the initialization first, in the
period until the role manager is started completely, the row
contains those null columns and any loging attempt in this period
will cause a memory access violation since those columns are not
expected to ever be null. This patch removes the race by starting
the password authenticator and autorizer only after the role manger
finished starting.

Tests:
  1. Unit tests (release)
  2. Auth and cqlsh dtests.

Fixes #4226