storage proxy: migrate to per scheduling group statistics
This commit build on top of the introduced per scheduling group
statistics template and employs it for achieving a per scheduling
group statistics in storage_proxy.

Some of the statistics also had meaning as a global - per
shard one. Those are the ones for determining if to
throtle the write request. This was handled by creating a
global stats struct that will hold those stats and by changing
the stat update to also include the global one.

One point that complicated it is an already existing aggregation
over the per shard stats that now became a per scheduling group
per shard stats, converting the aggregation to a two dimentional
aggregation.

one thing this commit doesn't handle is validating that an individual
statistic didn't "cross a scheduling group boundry", such validation
is possible but it can easily be added in the future. There is a
subtelty to doing so since if the operation did cross to other
scheduling groupm two connected statistics can loose balance
for example written bytes and completed write transactions.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>