imr: serializers: specialize structure_sizer, structure_serializer empty cases (#4)
structure_sizer and structure_serializer are both recursive templates.
Apart from the Continuation template parameter, they both accept
a Members template parameter which they process one by one by matching
it against a <typename Member, typename... Members> pattern for the
non-empty case.

The empty case is matched by the non-specialized case. However, this
is confusing, since it looks like the general case is ignoring its
Members... parameter. In fact it is correct since the only case where
the general case matches is when that parameter is empty.

To reduce confusion, eliminate the general case and instead provide
a specialization for the empty case.