imr: drop templated size_of completely and move to overloaded form
Now that the imr infrastructure expects the overloaded form of
size_of instead of the templated form, we can drop the templated
form completely. This will give us the extra protection of not
defaulting to some generic template form when the specific overloaded
form of size_of is not implemented.
After this transition, if a matching size_of overload is not
implemented when needed, the compiler will emit constraint not met
error.